+++ src/actor.rs
@@ -26,7 +26,7 @@ use utils::TimerFunc;
 ///
 /// Actors communicate exclusively by exchanging messages. Sender actor can
 /// wait for response. Actors are not referenced directly, but by
-/// non thread safe [Address<A>](struct.Address.html) or thread safe address
+/// non thread safe [`LocalAddress<A>`](struct.LocalAddress.html) or thread safe address
 /// [`SyncAddress<A>`](struct.SyncAddress.html)
 /// To be able to handle specific message actor has to provide
 /// [`Handler<M>`](trait.Handler.html)
@@ -105,7 +105,7 @@ pub trait Actor: Sized + 'static {
     ///     type Context = Context<Self>;
     /// }
     ///
-    /// let addr: Address<_> = MyActor.start();
+    /// let addr: LocalAddress<_> = MyActor.start();
     /// ```
     fn start<Addr>(self) -> Addr
         where Self: Actor<Context=Context<Self>> + ActorAddress<Self, Addr>
@@ -138,7 +138,7 @@ pub trait Actor: Sized + 'static {
     ///     type Context = Context<Self>;
     /// }
     ///
-    /// let addr: Address<_> = MyActor::create(|ctx: &mut Context<MyActor>| {
+    /// let addr: LocalAddress<_> = MyActor::create(|ctx: &mut Context<MyActor>| {
     ///     MyActor{val: 10}
     /// });
     /// ```
@@ -331,7 +331,7 @@ pub trait AsyncContext<A>: ActorContext + ToEnvelope<A> where A: Actor<Context=S
     /// }
     /// # fn main() {
     /// #    let sys = System::new("example");
-    /// #    let addr: Address<_> = MyActor.start();
+    /// #    let addr: LocalAddress<_> = MyActor.start();
     /// #    sys.run();
     /// # }
     /// ```
@@ -385,7 +385,7 @@ pub trait AsyncContext<A>: ActorContext + ToEnvelope<A> where A: Actor<Context=S
     /// }
     /// # fn main() {
     /// #    let sys = System::new("example");
-    /// #    let addr: Address<_> = MyActor.start();
+    /// #    let addr: LocalAddress<_> = MyActor.start();
     /// #    sys.run();
     /// # }
     /// ```
@@ -433,7 +433,7 @@ pub trait AsyncContext<A>: ActorContext + ToEnvelope<A> where A: Actor<Context=S
     /// }
     /// # fn main() {
     /// #    let sys = System::new("example");
-    /// #    let addr: Address<_> = MyActor.start();
+    /// #    let addr: LocalAddress<_> = MyActor.start();
     /// #    sys.run();
     /// # }
     /// ```