+++ CHANGES.md
@@ -6,6 +6,8 @@
 
 * Make rules of actor stopping more strict
 
+* Rename Address to a LocalAddress, SyncAddress to a Address
+
 
 ## 0.4.5 (2018-01-23)